STOR-10 move Exceptions into its own folder
